+++ R/splitdf.R
@@ -9,17 +9,23 @@
 #' # load German credit data
 #' data(germancredit)
 #'
+#' # Example I
 #' dt_list = split_df(germancredit, y="creditability")
-#' train = dt_list$train
-#' test = dt_list$test
+#' train = dt_list[[1]]
+#' test = dt_list[[2]]
 #'
 #' # dimensions of train and test datasets
 #' lapply(dt_list, dim)
 #'
+#'
+#' # Example II
+#' dt_list2 = split_df(germancredit, y="creditability", ratio = c(0.5, 0.2))
+#' lapply(dt_list2, dim)
+#'
 #' @import data.table
 #' @export
 split_df = function(dt, y=NULL, ratio=0.7, seed=186) {
-  rt = rn_train = rn_test = NULL
+  ind = NULL
 
   # set dt as data.table
   dt = setDT(dt)
@@ -29,36 +35,25 @@ split_df = function(dt, y=NULL, ratio=0.7, seed=186) {
   dt = rep_blank_na(dt)
 
   # set ratio range
-  if (!is.numeric(ratio) || length(ratio) != 1 || sum(ratio)>=1) {
+  if (!is.numeric(ratio) || length(ratio) >2 || sum(ratio)>1) {
     warning("Incorrect inputs; ratio must be a numeric that length equal to 1 and less than 1. It was set to 0.7.")
-    ratio = 0.7
-  }
-  # set seed
-  set.seed(seed)
-
-  rt = list(train=NULL, test=NULL)
-  if (is.null(y)) {
-    rn_sel = sample(nrow(dt), round(nrow(dt)*ratio))
-
-    rn_train = rn_sel
-    rn_test = setdiff(1:nrow(dt), rn_sel)
+    ratio = c(0.7, 0.3)
   } else {
-    # dt$y = dt[[y]]; dt[[y]] = NULL
-    y_unique = table(dt[[y]])
+    ratio_ = 1-sum(ratio)
+    if (ratio_ > 0) ratio = c(ratio, ratio_)
+  }
 
-    for (i in names(y_unique)) {
-      # dti = dt[.(i)]
-      rn_dti = which(dt[[y]]==i)
-      rn_sel = sample(rn_dti, round(length(rn_dti)*ratio))
-
-      rn_train = c(rn_train, rn_sel)
-      rn_test  = c(rn_test, setdiff(rn_dti, rn_sel))
-    }
+  # set seed and partition
+  set.seed(seed)
+  if (is.null(y)) {
+    dt[, ind := sample(length(ratio), .N, replace=TRUE, prob=ratio)]
+  } else {
+    dt[, ind := sample(length(ratio), .N, replace=TRUE, prob=ratio), by=y]
   }
 
   # random sort
-  rt$train = dt[sample(rn_train)]
-  rt$test  = dt[sample(rn_test)]
-
+  rt = list()
+  rt$train = dt[ind == 1,][, ind := NULL]
+  rt$test  = dt[ind == 2,][, ind := NULL]
   return(rt)
 }